requirements.txt
# Azure AI Vision SDK
azure-ai-vision-imageanalysis>=1.0.0b3
azure-core>=1.29.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Data handling
numpy>=1.24.0

# Progress bars and utilities
tqdm>=4.66.0

# Plotting and visualization
matplotlib>=3.7.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0 